pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["src/symfluence"]

[project]
name = "symfluence"
version = "0.7.0"        # <- static to avoid plugin issues
description = "SYnergistic Modelling Framework for Linking and Unifying Earth-system Nexii for Computational Exploration"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
authors = [{ name = "SYMFLUENCE Team", email = "dev@symfluence.org" }]

# Start with an empty list to unblock; we’ll fill runtime deps later.
dependencies = [
  "numpy",
  "pandas",
  "scipy",
  "xarray",
  "pint_xarray",
  "distributed",
  "pyviscous",
  "torch",
  "scikit-learn",
  "gdal",
  "geopandas",
  "rasterio",
  "pyproj",
  "shapely",
  "fiona",
  "easymore",
  "rasterstats",
  "pvlib",
  "cdo",
  "seaborn",
  "plotly",
  "netCDF4",
  "hydrobm",
  "baseflow",
  "matplotlib",
  "SALib",
  "psutil",
  "tqdm",
  "contextily",
  "pyyaml",
  "requests",
  "openai>=1.0.0",
  "cdsapi",
  "s3fs",
  "numexpr",
  "cftime",
  "pytest", 
  "sphinx", 
  "jupyterlab", 
  "ipykernel"
]

[project.scripts]
# thin wrapper that delegates to your existing CLI
symfluence = "symfluence.main_cli:main"

[project.urls]
Homepage = "https://symfluence.org"
Source = "https://github.com/DarriEy/SYMFLUENCE"
Issues = "https://github.com/DarriEy/SYMFLUENCE/issues"
Documentation = "https://symfluence.org/"

[tool.hatch.build.targets.wheel]
packages = ["src/symfluence"]
# Note: package data in src/symfluence/data/ is automatically included

[tool.hatch.build.targets.sdist]
include = [
  "src/symfluence",
  "README.md",
  "LICENSE",
  "MANIFEST.in",
  "requirements.txt",
  "examples/"
]

# NOTE: pytest configuration is in pytest.ini (authoritative source)
# This avoids duplication and ensures consistent test configuration